# doc-forensics-github/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Document Forensics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DocumentAnalysisInput {
  /** Type of document: aadhaar, pan, driving_license, degree, income_certificate, caste_certificate */
  documentType: string;
  /** Base64 encoded image of the document */
  imageBase64: string;
  /** Original filename */
  filename?: string;
}

export interface Finding {
  /** Category of finding (e.g. Font Inconsistency, Visual Artifact) */
  category: string;
  /** info, warning, or danger */
  severity: string;
  /** Detailed description of the finding */
  description: string;
}

export interface AnalysisResult {
  id: number;
  documentType: string;
  /** @nullable */
  filename?: string | null;
  /** authentic, suspicious, or tampered */
  verdict: string;
  /** 0-100 confidence score */
  confidence: number;
  /** 0-100 tampering score */
  tampering_score: number;
  findings: Finding[];
  summary: string;
  createdAt: string;
}

export interface AnalysisSummary {
  id: number;
  documentType: string;
  /** @nullable */
  filename?: string | null;
  verdict: string;
  confidence: number;
  tampering_score: number;
  summary: string;
  createdAt: string;
}

export interface ForensicsStats {
  total: number;
  authentic: number;
  suspicious: number;
  tampered: number;
  avgConfidence: number;
  avgTamperingScore: number;
}