# doc-forensics-github/lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Document Forensics API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummary {
  id: number;
  documentType: string;
  /** @nullable */
  filename?: string | null;
  verdict: string;
  confidence: number;
  tampering_score: number;
  summary: string;
  createdAt: string;
}