# doc-forensics-github/lib/api-zod/src/generated/types/documentAnalysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Document Forensics API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentAnalysisInput {
  /** Type of document: aadhaar, pan, driving_license, degree, income_certificate, caste_certificate */
  documentType: string;
  /** Base64 encoded image of the document */
  imageBase64: string;
  /** Original filename */
  filename?: string;
}