# doc-forensics-github/lib/api-zod/src/generated/types/forensicsStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Document Forensics API
 * OpenAPI spec version: 0.1.0
 */

export interface ForensicsStats {
  total: number;
  authentic: number;
  suspicious: number;
  tampered: number;
  avgConfidence: number;
  avgTamperingScore: number;
}